split plugins and put template in .cc
